first working path controller
- single target goal
- simple velocity sensing in joint state observer
- velocity PID tesed and fixed

+++ Src/stm32f1xx_it.c
@@ -210,7 +210,7 @@ static void averageAndStoreAdcHalfSamples(size_t startIndex) {
 	}
 	
 	for (size_t channel = 0; channel < ADC_CHANNELS_IN_USE; channel++) {
-		gAdcCurrentValue[channel] = sum[channel] / (ADC_DMA_BUF_NUM_SAMPLES / 2);
+		gAdcCurrentValue[channel] = (float)sum[channel] / (float)(ADC_DMA_BUF_NUM_SAMPLES / 2);
 	}
 }
